Restructuring for maintainability
