+++ src/oss4energy/scripts/__init__.py
@@ -3,14 +3,12 @@ Module containing methods to be run in scripts
 """
 
 import os
-from ftplib import FTP
 from urllib.request import urlretrieve
 
 import pandas as pd
 import tomllib
 from tomlkit import document, dump
 
-from oss4energy.src.config import SETTINGS
 from oss4energy.src.helpers import sorted_list_of_unique_elements
 from oss4energy.src.log import log_info
 from oss4energy.src.nlp.markdown_io import markdown_to_clean_plaintext
@@ -263,46 +261,6 @@ def generate_listing(target_output_file: str = FILE_OUTPUT_LISTING_CSV) -> None:
     """
     )
     format_files()
-
-
-def publish_to_ftp() -> None:
-    """Exports data generated to FTP (requires .env to be defined with credentials to the FTP)
-
-    :raises EnvironmentError: when the FTP credentials are not given in environment
-    """
-    for i in [
-        SETTINGS.EXPORT_FTP_URL,
-        SETTINGS.EXPORT_FTP_USER,
-        SETTINGS.EXPORT_FTP_PASSWORD,
-    ]:
-        if i is None:
-            raise EnvironmentError(
-                f"{i.__name__} must be defined for FTP export to work"
-            )
-        if len(i) == 0:
-            raise EnvironmentError(
-                f"{i.__name__} must have an adequate value for FTP export to work"
-            )
-    files_out = [
-        FILE_OUTPUT_SUMMARY_TOML,
-        FILE_OUTPUT_LISTING_CSV,
-        FILE_OUTPUT_LISTING_FEATHER,
-    ]
-
-    with FTP(
-        host=SETTINGS.EXPORT_FTP_URL,
-        user=SETTINGS.EXPORT_FTP_USER,
-        passwd=SETTINGS.EXPORT_FTP_PASSWORD,
-    ) as ftp:
-        try:
-            ftp.mkd("oss4energy")
-        except:
-            pass
-        ftp.cwd("oss4energy")
-        for i in files_out:
-            with open(i, "rb") as fp:
-                log_info(f"Uploading {i}")
-                ftp.storbinary("STOR %s" % os.path.basename(i), fp, blocksize=1024)
 
 
 def search_in_listing() -> None:
